# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Runtime Platform API
  version: 0.1.0
  description: Multi-tenant resource API. Every operation is authorized against the tenant and application claims in the OIDC access token.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - oidc: []
...
paths:
...
  /v1/approvals:
    get:
      summary: List pending approvals visible to the current application
      tags: [Approvals]
      operationId: listPendingApprovals
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending]
            default: pending
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Pending approvals ordered from oldest to newest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/approvals/{approvalId}:decide:
    post:
      summary: Decide a pending approval
      tags: [Approvals]
      operationId: decideApproval
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Approval decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Version is stale or approval is no longer pending
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://iam.example.com/.well-known/openid-configuration
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The caller lacks permission for this tenant resource
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource is absent or not visible to the current tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [version, decision]
      properties:
        version:
          type: integer
          minimum: 1
        decision:
          type: string
          enum: [allow_once, allow_session, deny]
        reason:
          type: string
          maxLength: 1000
    ApprovalListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalSummary'
    ApprovalSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - run_id
        - version
        - status
        - workspace_name
        - agent_name
        - tool_name
        - tool_call_id
        - effect
        - sandbox
        - binding_digest
        - arguments
        - policy_digest
        - session_scope_digest
        - policy_snapshot
        - available_decisions
        - created_at
      properties:
        id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [pending]
        workspace_name:
          type: string
        agent_name:
          type: string
        tool_name:
          type: string
        tool_call_id:
          type: string
        effect:
          type: string
          enum: [pure, idempotent, non_idempotent, unknown]
        sandbox:
          type: string
          enum: [restricted_container, kata, trusted_native]
        binding_digest:
          type: string
          pattern: '^[0-9a-f]{64}$'
        arguments:
          type: object
          additionalProperties: true
        policy_digest:
          anyOf:
            - type: string
              pattern: '^[0-9a-f]{64}$'
            - type: 'null'
        session_scope_digest:
          anyOf:
            - type: string
              pattern: '^[0-9a-f]{64}$'
            - type: 'null'
        policy_snapshot:
          anyOf:
            - $ref: '#/components/schemas/ApprovalPolicySnapshot'
            - type: 'null'
        available_decisions:
          type: array
          minItems: 2
          uniqueItems: true
          items:
            type: string
            enum: [allow_once, allow_session, deny]
        created_at:
          type: string
          format: date-time
    ApprovalPolicySnapshot:
      type: object
      additionalProperties: false
      required:
        - approval
        - effect
        - implementation_digest
        - required_scopes
        - sandbox
        - tool_name
      properties:
        approval:
          type: string
          const: ask
        effect:
          type: string
          enum: [pure, idempotent, non_idempotent, unknown]
        implementation_digest:
          type: string
          pattern: '^[0-9a-f]{64}$'
        required_scopes:
          type: array
          uniqueItems: true
          items:
            type: string
        sandbox:
          type: string
          enum: [restricted_container, kata, trusted_native]
        tool_name:
          type: string
    Approval:
      type: object
      additionalProperties: false
      required: [id, tenantId, runId, version, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        version:
          type: integer
        status:
          type: string
          enum: [pending, approved, denied, expired]
        createdAt:
          type: string
          format: date-time
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string